conti refactor ec2
